+++ src/components/employees/employeeList.js
@@ -3,6 +3,7 @@
 var React = require('react');
 var Router = require('react-router');
 var Link = Router.Link;
+var swal = require('sweetalert2');
 var EmployeeActions = require('../../actions/employeeActions');
 
 var EmployeeList = React.createClass({
@@ -12,7 +13,21 @@ var EmployeeList = React.createClass({
 
 	deleteEmployee: function(id, event) {
 		event.preventDefault();
-		EmployeeActions.deleteEmployee(id);
+		var _this = this;
+		swal({
+			title: '',
+			text: 'Are you sure you want to delete this employee?',
+			type: 'warning',
+			showCancelButton: true,
+			allowOutsideClick: false,
+			allowEscapeKey: false,
+			allowEnterKey: false
+		}).then(function (result) {
+			if (result.value) {
+				_this.props.onAction();
+				EmployeeActions.deleteEmployee(id);
+			}
+		});
 	},
 
 	render: function() {
@@ -22,8 +37,20 @@ var EmployeeList = React.createClass({
 					<td><Link to="manageEmployee" params={{id: employee.id}}>{employee.id}</Link></td>
 					<td>{employee.fullName}</td>
 					<td className="text-center action">
-						<Link to="manageEmployee" params={{ id: employee.id }}>Edit</Link>
-						<a href="#" onClick={this.deleteEmployee.bind(this, employee.id)}>Delete</a>
+						<Link to="manageEmployee" title="Edit"
+							className="btn btn-default action-button"
+							params={{ id: employee.id }}
+							disabled={this.props.loader}>
+
+							<i className="glyphicon glyphicon-pencil"></i>
+						</Link>
+
+						<button className="btn btn-default action-button delete-btn"
+							onClick={this.deleteEmployee.bind(this, employee.id)}
+							disabled={this.props.loader} title="Delete">
+							
+							<i className="glyphicon glyphicon-trash"></i>
+						</button>
 					</td>
 				</tr>
 			);
